// src/router/enum.ts
export enum ROUTE_MAP {
  /**
   * 首页
   */
  app = 'app',
  /**
   * Loading
   */
  loading = 'loading',
  /**
   * Modal
   */
  modal = 'modal',
  /**
   * Message
   */
  message = 'message',
  /**
   * Button
   */
  button = 'button',
  /**
   * Switch
   */
  switch = 'switch',
  /**
   * Radio
   */
  radio = 'radio',
  /**
   * Checkbox
   */
  checkbox = 'checkbox',
  /**
   * Input
   */
  input = 'input',
  /**
   * Transition
   */
  transition = 'transition'
}